refactor: make ULA a 16 bit (because the PC)

+++ src/ula.c
@@ -8,7 +8,7 @@
 #include "main.h"
 
 
-ULAOut ula(const int8_t input1, const int8_t input2, const int ulaControl) {
+ULAOut ula(const int16_t input1, const int16_t input2, const int ulaControl) {
     ULAOut out;
     // Garante que os valores de UlaOUT comecem com valor limpo
     out.value = 0;
@@ -23,24 +23,24 @@ ULAOut ula(const int8_t input1, const int8_t input2, const int ulaControl) {
         case 1: // ADDI
         case 3: // LW. Calcula endereço somando os regs
         case 7: // SW. Calcula endereço somando os regs
-            out.value = (int8_t) (input1 + input2);
+            out.value = (int16_t) (input1 + input2);
             // Check if overflow was occurred with positive
             out.overflow = (input1 > 0 && input2 > 0 && out.value < 0) || (input1 < 0 && input2 < 0 && out.value > 0);
             // se resultar em overflow, manda sinal para out.overflow
             break;
         case 2: // SUB/BEQ
         case 6:
-            out.value = (int8_t) (input1 - input2);
+            out.value = (int16_t) (input1 - input2);
             // Check if overflow was occurred with negative
             out.overflow = (input1 > 0 && input2 < 0 && out.value < 0) || (input1 < 0 && input2 > 0 && out.value > 0);
             // Util only for branch (ulaControl = 6), check if sub of inputs is 0 (two equal inputs subtracted is 0)
             out.equal = (out.value == 0);
             break;
         case 4: // AND
-            out.value = (int8_t) (input1 & input2);
+            out.value = (int16_t) (input1 & input2);
             break;
         case 5: // OR
-            out.value = (int8_t) (input1 | input2);
+            out.value = (int16_t) (input1 | input2);
             break;
         default:
             break;
